# api/src/handlers/archive.rs
//! Handler for the archive endpoint: streams `git archive` output.
//!
//! Archives are generated by the git binary, not libgit2 (docs/DECISIONS.md
//! #2 hybrid approach). Only the resolved full sha is passed to the exec, so
//! no user-controlled string ever reaches the command line.

use std::process::Stdio;

use axum::body::Body;
use axum::extract::{Path, State};
use axum::http::{HeaderMap, header};
use axum::response::Response;
use tokio::io::AsyncReadExt;
use tokio::process::Command;
use tokio_util::io::ReaderStream;

use super::{
    IMMUTABLE_CACHE_CONTROL, NO_CACHE_CONTROL, if_none_match, not_modified, validator_etag,
};
use crate::error::{ApiError, ErrorResponse};
use crate::repo::{meta, open, resolve};
use crate::state::AppState;

#[derive(Debug, Clone, Copy, PartialEq, Eq)]
enum ArchiveFormat {
    TarGz,
    Zip,
}

impl ArchiveFormat {
    /// Value for `git archive --format` (which gzips `tar.gz` itself).
    fn format_arg(self) -> &'static str {
        match self {
            Self::TarGz => "tar.gz",
            Self::Zip => "zip",
        }
    }

    fn content_type(self) -> &'static str {
        match self {
            Self::TarGz => "application/gzip",
            Self::Zip => "application/zip",
        }
    }

    fn extension(self) -> &'static str {
        self.format_arg()
    }
}

/// Splits the `{ref}.{format}` wildcard. The ref may contain `/` and dots, so
/// only the known format suffixes are matched (`.tar.gz` before `.zip`, which
/// also disambiguates a hypothetical ref ending in `.zip`).
fn parse_archive_target(rest: &str) -> Result<(&str, ArchiveFormat), ApiError> {
    let (refname, format) = if let Some(refname) = rest.strip_suffix(".tar.gz") {
        (refname, ArchiveFormat::TarGz)
    } else if let Some(refname) = rest.strip_suffix(".zip") {
        (refname, ArchiveFormat::Zip)
    } else {
        return Err(ApiError::InvalidParam(format!(
            "unsupported archive format in '{rest}' (expected .tar.gz or .zip)"
        )));
    };
    if refname.is_empty() {
        return Err(ApiError::InvalidParam(format!(
            "missing ref in archive path '{rest}'"
        )));
    }
    Ok((refname, format))
}

/// Replaces everything outside `[A-Za-z0-9._-]` with `-` (`feature/x` →
/// `feature-x`). The result is ASCII-safe for the `Content-Disposition`
/// filename (no RFC 6266 escaping needed) and the archive prefix.
fn sanitize_component(component: &str) -> String {
    component
        .chars()
        .map(|c| match c {
            'A'..='Z' | 'a'..='z' | '0'..='9' | '.' | '_' | '-' => c,
            _ => '-',
        })
        .collect()
}

/// Source archive
///
/// Streams `git archive` output chunked (no `Content-Length`). The ref is
/// resolved first and only the **full sha** reaches the command line. The
/// `{ref}.{format}` split is suffix matching: `.tar.gz` first, then `.zip`.
#[utoipa::path(
    get,
    path = "/api/v1/repos/{repo}/archive/{ref}.{format}",
    tag = "files",
    params(
        ("repo" = String, Path, description = "Repository name without the `.git` suffix", example = "git-compose"),
        ("ref" = String, Path, description = "Branch, tag, or commit sha", example = "main"),
        ("format" = String, Path, description = "`tar.gz` or `zip`", example = "tar.gz"),
    ),
    responses(
        (status = 200, description = "Archive stream (`application/gzip` or `application/zip`). If `git archive` fails mid-stream the response is truncated without a status change.",
            content_type = "application/gzip",
            body = String,
            headers(
                ("Content-Disposition" = String, description = "`attachment; filename=\"{repo}-{safe_ref}.{format}\"`"),
                ("X-Content-Type-Options" = String, description = "Always `nosniff`"),
                ("ETag" = String, description = "**Weak** (`W/\"...\"`); absent on full-sha requests"),
                ("Cache-Control" = String, description = "`no-cache`, or `public, max-age=31536000, immutable` for a full sha"),
            ),
        ),
        (status = 304, description = "`If-None-Match` matched — no git process is spawned"),
        (status = 400, description = "`invalid_param` — suffix is neither `.tar.gz` nor `.zip`", body = ErrorResponse),
        (status = 404, description = "`repo_not_found`, `ref_not_found`", body = ErrorResponse),
        (status = 500, description = "`internal` — `git archive` could not be spawned", body = ErrorResponse),
    ),
)]
pub async fn get_archive(
    State(state): State<AppState>,
    Path((name, rest)): Path<(String, String)>,
    headers: HeaderMap,
) -> Result<Response, ApiError> {
    let (refname, format) = parse_archive_target(&rest)?;
    let refname = refname.to_owned();

    let root = state.config.repo_root.clone();
    let block_name = name.clone();
    let block_refname = refname.clone();
    let (git_dir, sha, immutable, validator) = tokio::task::spawn_blocking(move || {
        let repo = open::open_named(&root, &block_name)?;
        let validator = meta::validator(&repo);
        let commit = resolve::resolve_commit(&repo, &block_refname)?;
        let sha = commit.id().to_string();
        let immutable = block_refname == sha;
        Ok::<_, ApiError>((repo.path().to_path_buf(), sha, immutable, validator))
    })
    .await
    .map_err(|err| ApiError::Internal(err.into()))??;

    // Weak ETag: the archive bytes are not guaranteed identical across git
    // invocations, but the content is equivalent for an unchanged validator.
    // A match skips spawning `git archive` entirely.
    let etag = (!immutable).then(|| format!("W/{}", validator_etag(&validator)));
    if let Some(etag) = &etag
        && if_none_match(&headers, etag)
    {
        return Ok(not_modified(etag));
    }

    let base = format!(
        "{}-{}",
        sanitize_component(&name),
        sanitize_component(&refname)
    );
    let mut child = Command::new("git")
        .arg("--git-dir")
        .arg(&git_dir)
        .arg("archive")
        .arg("--format")
        .arg(format.format_arg())
        .arg("--prefix")
        .arg(format!("{base}/"))
        .arg(&sha)
        .stdin(Stdio::null())
        .stdout(Stdio::piped())
        .stderr(Stdio::piped())
        .kill_on_drop(true)
        .spawn()
        .map_err(|err| {
            ApiError::Internal(anyhow::Error::new(err).context("failed to spawn git archive"))
        })?;

    let stdout = child.stdout.take().expect("stdout was configured as piped");
    let mut stderr = child.stderr.take().expect("stderr was configured as piped");

    // Reaper: collects stderr and waits so the child never becomes a zombie.
    // Failures past this point cannot change the status code (headers are
    // already sent); the stream just ends short and the client sees a broken
    // download. A dropped Body closes the pipe and git exits on EPIPE.
    tokio::spawn(async move {
        let mut diagnostics = Vec::new();
        let _ = stderr.read_to_end(&mut diagnostics).await;
        match child.wait().await {
            Ok(status) if status.success() => {}
            outcome => tracing::error!(
                sha,
                ?outcome,
                stderr = %String::from_utf8_lossy(&diagnostics),
                "git archive exited abnormally"
            ),
        }
    });

    let mut response = Response::builder()
        .header(header::CONTENT_TYPE, format.content_type())
        .header(
            header::CONTENT_DISPOSITION,
            format!("attachment; filename=\"{base}.{}\"", format.extension()),
        )
        // Repository contents are untrusted input; never let the browser sniff.
        .header(header::X_CONTENT_TYPE_OPTIONS, "nosniff");
    response = match &etag {
        None => response.header(header::CACHE_CONTROL, IMMUTABLE_CACHE_CONTROL),
        Some(etag) => response
            .header(header::ETAG, etag)
            .header(header::CACHE_CONTROL, NO_CACHE_CONTROL),
    };
    response
        .body(Body::from_stream(ReaderStream::new(stdout)))
        .map_err(|err| ApiError::Internal(err.into()))
}

#[cfg(test)]
mod tests {
    use super::*;

    #[test]
    fn parse_archive_target_should_split_known_suffixes() {
        let (refname, format) = parse_archive_target("main.tar.gz").unwrap();
        assert_eq!(refname, "main");
        assert_eq!(format, ArchiveFormat::TarGz);

        let (refname, format) = parse_archive_target("feature/x.zip").unwrap();
        assert_eq!(refname, "feature/x");
        assert_eq!(format, ArchiveFormat::Zip);

        // `.tar.gz` wins over a ref that happens to end in `.gz`.
        let (refname, format) = parse_archive_target("v1.0.tar.gz").unwrap();
        assert_eq!(refname, "v1.0");
        assert_eq!(format, ArchiveFormat::TarGz);
    }

    #[test]
    fn parse_archive_target_should_reject_unknown_or_empty() {
        for rest in ["main", "main.tar", "main.rar", ".tar.gz", ".zip"] {
            assert!(
                matches!(parse_archive_target(rest), Err(ApiError::InvalidParam(_))),
                "rest {rest:?} was not rejected"
            );
        }
    }

    #[test]
    fn sanitize_component_should_keep_safe_chars_only() {
        assert_eq!(sanitize_component("feature/x"), "feature-x");
        assert_eq!(sanitize_component("v1.0_rc-2"), "v1.0_rc-2");
        assert_eq!(sanitize_component("한글 \"quote\""), "----quote-");
    }
}