# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Axgit API",
    "description": "Read-only JSON API over bare Git repositories, plus Smart HTTP (fetch/clone) and an Atom feed.\n\nThere are no mutation endpoints and no authentication: writes happen over SSH on the git-server container, never here.\n\nThis document describes shapes, parameters and status codes. The semantic rules it cannot express — truncation caps, longest-ref matching for the `{ref}`/`{path}` split, merge-commit log simplification, and the exact conditions under which a field is `null` — live in `docs/API.md`, which remains the normative contract.",
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/repos/{repo}/archive/{ref}.{format}": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Source archive",
        "description": "Streams `git archive` output chunked (no `Content-Length`). The ref is\nresolved first and only the **full sha** reaches the command line. The\n`{ref}.{format}` split is suffix matching: `.tar.gz` first, then `.zip`.",
        "operationId": "get_archive",
        "parameters": [
          {
            "name": "repo",
            "in": "path",
            "description": "Repository name without the `.git` suffix",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "git-compose"
          },
          {
            "name": "ref",
            "in": "path",
            "description": "Branch, tag, or commit sha",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "main"
          },
          {
            "name": "format",
            "in": "path",
            "description": "`tar.gz` or `zip`",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "tar.gz"
          }
        ],
        "responses": {
          "200": {
            "description": "Archive stream (`application/gzip` or `application/zip`). If `git archive` fails mid-stream the response is truncated without a status change.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "`no-cache`, or `public, max-age=31536000, immutable` for a full sha"
              },
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "description": "`attachment; filename=\"{repo}-{safe_ref}.{format}\"`"
              },
              "ETag": {
                "schema": {
                  "type": "string"
                },
                "description": "**Weak** (`W/\"...\"`); absent on full-sha requests"
              },
              "X-Content-Type-Options": {
                "schema": {
                  "type": "string"
                },
                "description": "Always `nosniff`"
              }
            },
            "content": {
              "application/gzip": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "`If-None-Match` matched — no git process is spawned"
          },
          "400": {
            "description": "`invalid_param` — suffix is neither `.tar.gz` nor `.zip`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`repo_not_found`, `ref_not_found`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "`internal` — `git archive` could not be spawned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "One of `repo_not_found`, `ref_not_found`, `path_not_found`,\n`invalid_param`, `read_only`, `internal`.",
            "example": "repo_not_found"
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. Internal errors report a generic message; the\ncause only goes to the server log.",
            "example": "repository 'foo' not found"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Error envelope shared by every failing endpoint (docs/API.md).",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
    }
  },
...
}